3. inner class to upper level
